+++ R/simulation-simInit.R
@@ -360,20 +360,29 @@ setMethod(
     names(modules) <- unlist(modules)
 
     # Check that modules exist in paths$modulePath
-    moduleDirsPoss <- lapply(modules, function(m) file.path(sim@paths$modulePath, m))
-    moduleDirsExist <- lapply(moduleDirsPoss, function(poss) dir.exists(poss))
-    # dir.exists(file.path(paths$modulePath, unlist(modules)))
-    if (!isTRUE(all(unlist(lapply(moduleDirsExist, any))))) {
-      stop("These modules: ", unlist(modules)[!moduleDirsExist], " , don't exist in ", paths$modulePath)
-    }
-    modulePaths <- Map(poss = moduleDirsPoss, exist = moduleDirsExist, function(poss, exist)
-      poss[exist][1])
+    modulePaths <- .checkModuleDirsAndFiles(modules = modules, modulePath = sim@paths$modulePath)
 
     # identify childModules, recursively
     childModules <- .identifyChildModules(sim = sim, modules = modulePaths)
     modules <- as.list(unique(unlist(childModules))) # flat list of all modules
     names(modules) <- unlist(modules)
     modules <- lapply(modules, basename2)
+
+    moduleNames <- names(modules)
+    names(moduleNames) <- moduleNames
+
+    # Test that all child module files and dirs exist
+    childModDirsExist <- lapply(moduleNames, dir.exists)
+    childModMainFiles <- file.path(moduleNames, paste0(basename2(moduleNames), ".R"))
+    names(childModMainFiles) <- basename2(moduleNames)
+    childModFilesExist <- file.exists(childModMainFiles)
+    if (any(!childModFilesExist)) {
+      stop(childModMainFiles[!childModFilesExist], " does not exist; please create one or diagnose. ",
+           "  Some possible causes:\n  ",
+           "- git submodule not initiated?\n  ",
+           "- the module was not created using 'newModule(...)' so is missing key files")
+    }
+
 
     modules <- modules[!sapply(modules, is.null)] %>%
       lapply(., `attributes<-`, list(parsed = FALSE))
@@ -993,23 +1002,26 @@ simInitAndExperiment <- function(times, params, modules, objects, paths, inputs,
     message("Duplicate module, ", modules[duplicated(modules)], ", specified. Skipping loading it twice.")
   }
   if (length(modules) > 0) {
-    modulesToSearch <- lapply(.parseModulePartial(sim, modulesToSearch,
+    modulesToSearch3 <- lapply(.parseModulePartial(sim, modulesToSearch,
                                                   defineModuleElement = "childModules",
                                                   envir = sim@.xData[[".parsedFiles"]]),
                               as.list)
-    isParent <- unlist(lapply(modulesToSearch, function(x) length(x)>1))
+    if (length(modulesToSearch3) > 0) {
+      isParent <- unlist(lapply(modulesToSearch3, function(x) length(x)>0))
 
-    modulesToSearch[isParent] <- Map(x = modulesToSearch[isParent],
-                                     nam = dirname(names(modulesToSearch[isParent])),
-                                     function(x, nam){
-                                       mods <- lapply(x, function(y) file.path(nam, y))
-                                       names(mods) <- unlist(lapply(mods, basename2))
+      modulesToSearch3[isParent] <- Map(x = modulesToSearch3[isParent],
+                                       nam = dirname(names(modulesToSearch3[isParent])),
+                                       function(x, nam){
+                                         mods <- lapply(x, function(y) file.path(nam, y))
+                                         names(mods) <- unlist(lapply(mods, basename2))
 
-                                       .identifyChildModules(sim = sim, modules = mods)}
-    )
-    modulesToSearch2 <- as.list(names(modulesToSearch[!isParent]))
-    names(modulesToSearch2) <- names(modulesToSearch[!isParent])
-    modulesToSearch[!isParent] <- modulesToSearch2
+                                         .identifyChildModules(sim = sim, modules = mods)}
+      )
+      modulesToSearch2 <- as.list(names(modulesToSearch3[!isParent]))
+      names(modulesToSearch2) <- names(modulesToSearch3[!isParent])
+      modulesToSearch3[!isParent] <- modulesToSearch2
+      modulesToSearch <- modulesToSearch3
+    }
   }
   return(modulesToSearch)
 }
@@ -1184,4 +1196,30 @@ simInitAndExperiment <- function(times, params, modules, objects, paths, inputs,
   # if (isTRUE(isMissing["loadOrder"])) li$loadOrder <- .loadOrderDefault
 
   return(li)
+}
+
+.checkModuleDirsAndFiles <- function(modules, modulePath) {
+  moduleDirsPoss <- lapply(modules, function(m) file.path(modulePath, m))
+  moduleDirsExist <- lapply(moduleDirsPoss, function(poss) dir.exists(poss))
+  moduleFilesPoss <- lapply(moduleDirsPoss, function(poss) file.path(file.path(poss,
+                                                                               paste0(basename(poss), ".R"))))
+  moduleFilesExist <- lapply(moduleFilesPoss, function(poss) file.exists(poss))
+  # dir.exists(file.path(paths$modulePath, unlist(modules)))
+  if (!isTRUE(all(unlist(lapply(moduleDirsExist, any))))) {
+    if (sum(!unlist(moduleDirsExist)) > 1) {
+      moduleTxt1 <- "These modules"
+      moduleTxt2 <- "don't"
+    } else {
+      moduleTxt1 <- "This module"
+      moduleTxt2 <- "doesn't"
+    }
+    stop(moduleTxt1, ":\n    ", unlist(modules)[!unlist(moduleDirsExist)], ",\n  ",
+         moduleTxt2," exist in:\n    ", modulePath)
+  }
+  if (!isTRUE(all(unlist(lapply(moduleFilesExist, any))))) {
+    stop("These main module file(s) are missing:\n    ",
+         paste(unlist(moduleFilesPoss)[!unlist(moduleFilesExist)], sep = "\n  "))
+  }
+  modulePaths <- Map(poss = moduleDirsPoss, exist = moduleDirsExist, function(poss, exist)
+    poss[exist][1])
 }